@@ -18,26 +18,21 @@ use Webmozart\Assert\Assert;
 final class TableDecorator implements ClassDecorator
 {
 
-	/** @var string */
-	private $schema;
+	private string $schema;
 
-	/** @var string */
-	private $tableName;
+	private string $tableName;
 
-	/** @var string */
-	private $primaryKeyClass;
+	private string $primaryKeyClass;
 
-	/** @var string */
-	private $rowClass;
+	private string $rowClass;
 
-	/** @var string */
-	private $modificationClass;
+	private string $modificationClass;
 
 	/** @var Column[] */
-	private $columnInfo;
+	private array $columnInfo;
 
 	/** @var array<string, Type> */
-	private $columnPhpTypes;
+	private array $columnPhpTypes;
 
 	/**
 	 * @param array<string, Column> $columnInfo
@@ -270,10 +265,7 @@ final class TableDecorator implements ClassDecorator
 		$this->implementConfigMethod($classType, $name, new Code\PhpLiteral($namespace->unresolveName($class) . '::class'));
 	}
 
-	/**
-	 * @param mixed $value
-	 */
-	private function implementConfigMethod(Code\ClassType $classType, string $name, $value): void
+	private function implementConfigMethod(Code\ClassType $classType, string $name, mixed $value): void
 	{
 		$classType->addMethod($name)
 			->setStatic()
