@@ -7,21 +7,12 @@ namespace Grifart\Tables\Scaffolding;
 final class Column
 {
 
-	private string $name;
-
-	private string $type;
-
-	private bool $nullable;
-
-	private bool $hasDefaultValue;
-
-	public function __construct(string $name, string $type, bool $nullable, bool $hasDefaultValue)
-	{
-		$this->name = $name;
-		$this->type = $type;
-		$this->nullable = $nullable;
-		$this->hasDefaultValue = $hasDefaultValue;
-	}
+	public function __construct(
+		private string $name,
+		private string $type,
+		private bool $nullable,
+		private bool $hasDefaultValue,
+	) {}
 
 	public function getName(): string
 	{
